Fix the tests for iterators
